Move unit tests alongside source
