+++ src/wasm-bindings/cc_web_components.spec.ts
@@ -1,0 +1,8 @@
+import { describe, it, expect } from 'vitest';
+import { process_drag } from './cc_web_components.js';
+
+describe('process_drag', () => {
+    it('returns the delta value', () => {
+        expect(process_drag(5)).toBe(5);
+    });
+});
