+++ tsconfig.json
@@ -9,6 +9,6 @@
         "forceConsistentCasingInFileNames": true,
         "outDir": "dist/types"
     },
-    "include": ["src/**/*", "tests/**/*"],
+    "include": ["src/**/*"],
     "exclude": ["node_modules", "dist"]
 }
